finish chat APIs

+++ src/main/kotlin/com/moyeoba/moyeoba/data/dto/response/GetChatRoomsResponseDto.kt
@@ -1,6 +1,6 @@
 package com.moyeoba.moyeoba.data.dto.response
 
 data class GetChatRoomsResponseDto (
-    val roomIdList: List<Int>,
+    val roomIdList: List<Long>,
     val lastMsg: List<String>
 )
